+++ @planetarium/account-web3-secret-storage/test/MockPassphraseEntry.ts
@@ -2,12 +2,17 @@ import { KeyId } from "../src/KeyId";
 import { Passphrase, PassphraseEntry } from "../src/PassphraseEntry";
 
 export class MockPassphraseEntry implements PassphraseEntry {
+  rightPassphrase: Passphrase;
   authenticateCalls: { keyId: KeyId; firstAttempt: boolean }[] = [];
   configurePassphraseCalls: number = 0;
   temporaryAuthenticateResult: { passphrase: Passphrase; times: number } = {
     passphrase: "passphrase",
     times: 0,
   };
+
+  constructor(rightPassphrase: Passphrase = "passphrase") {
+    this.rightPassphrase = rightPassphrase;
+  }
 
   setTemporaryAuthenticateResult(passphrase: Passphrase, times: number) {
     this.temporaryAuthenticateResult = { passphrase, times };
@@ -19,11 +24,14 @@ export class MockPassphraseEntry implements PassphraseEntry {
       this.temporaryAuthenticateResult.times--;
       return this.temporaryAuthenticateResult.passphrase;
     }
-    return "passphrase";
+    if (this.authenticateCalls.length > 10) {
+      throw new Error("Failed to authenticate too many times; probably a bug");
+    }
+    return this.rightPassphrase;
   }
 
-  async configurePassphrase(): Promise<string> {
+  async configurePassphrase(): Promise<Passphrase> {
     this.configurePassphraseCalls++;
-    return "passphrase";
+    return this.rightPassphrase;
   }
 }